Refactored code to use a directive for shared functions in components
