+++ src/app/features/bulk-upload-v2/bulk-upload-references/staff-references/staff-references-page.component.ts
@@ -1,11 +1,10 @@
 import { I18nPluralPipe } from '@angular/common';
 import { HttpErrorResponse } from '@angular/common/http';
 import { Component, OnInit } from '@angular/core';
-import { FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
+import { FormBuilder } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
 import { JourneyType } from '@core/breadcrumb/breadcrumb.model';
-import { ErrorDefinition, ErrorDetails } from '@core/model/errorSummary.model';
-import { Establishment, LocalIdentifiersRequest } from '@core/model/establishment.model';
+import { Establishment } from '@core/model/establishment.model';
 import { URLStructure } from '@core/model/url.model';
 import { Worker } from '@core/model/worker.model';
 import { BackService } from '@core/services/back.service';
@@ -27,15 +26,8 @@ import { BulkUploadReferencesDirective } from '../bulk-upload-references.directi
   providers: [I18nPluralPipe],
 })
 export class StaffReferencesComponent extends BulkUploadReferencesDirective implements OnInit {
-  private maxLength = 50;
-  public form: FormGroup;
-  public references: Worker[] = [];
   private primaryWorkplace: Establishment;
   private subscriptions: Subscription = new Subscription();
-  public formErrorsMap: ErrorDetails[] = [];
-  public serverError: string;
-  public serverErrorsMap: ErrorDefinition[] = [];
-  public submitted = false;
   public return: URLStructure = { url: ['/dev', 'bulk-upload', 'workplace-references'] };
   private establishmentUid: string;
 
@@ -50,7 +42,7 @@ export class StaffReferencesComponent extends BulkUploadReferencesDirective impl
     private breadcrumbService: BreadcrumbService,
     private workerService: WorkerService,
   ) {
-    super(errorSummaryService);
+    super(errorSummaryService, formBuilder);
   }
 
   ngOnInit(): void {
@@ -67,56 +59,6 @@ export class StaffReferencesComponent extends BulkUploadReferencesDirective impl
     this.setServerErrors();
   }
 
-  protected setupForm(): void {
-    this.submitted = false;
-    this.form = this.formBuilder.group({}, { validator: this.checkDuplicates });
-    this.references.forEach((reference: Worker) => {
-      this.form.addControl(
-        `reference-${reference.uid}`,
-        new FormControl(reference.localIdentifier, [Validators.required, Validators.maxLength(this.maxLength)]),
-      );
-
-      this.formErrorsMap.push({
-        item: `reference-${reference.uid}`,
-        type: [
-          {
-            name: 'required',
-            message: `Enter the missing workplace reference.`,
-          },
-          {
-            name: 'maxlength',
-            message: `The reference must be ${this.maxLength} characters or less.`,
-          },
-          {
-            name: 'duplicate',
-            message: `Enter a different workplace reference.`,
-          },
-        ],
-      });
-    });
-  }
-
-  public onSubmit(event: Event): void {
-    event.preventDefault();
-    this.submitted = true;
-    this.errorSummaryService.syncFormErrorsEvent.next(true);
-
-    if (this.form.valid) {
-      this.save();
-    } else {
-      this.errorSummaryService.scrollToErrorSummary();
-    }
-  }
-
-  protected onError(response: HttpErrorResponse): void {
-    if (response.status === 400) {
-      this.serverErrorsMap[1].message += ` '${response.error.duplicateValue}' has previously been used.`;
-    }
-
-    this.serverError = this.errorSummaryService.getServerErrorMessage(response.status, this.serverErrorsMap);
-    this.errorSummaryService.scrollToErrorSummary();
-  }
-
   private setServerErrors() {
     this.serverErrorsMap = [
       {
@@ -125,7 +67,7 @@ export class StaffReferencesComponent extends BulkUploadReferencesDirective impl
       },
       {
         name: 400,
-        message: `Unable to update workplace reference.`,
+        message: `Unable to update staff reference.`,
       },
     ];
   }
@@ -142,23 +84,5 @@ export class StaffReferencesComponent extends BulkUploadReferencesDirective impl
           (error: HttpErrorResponse) => this.onError(error),
         ),
     );
-  }
-
-  public setReturn(): void {
-    this.bulkUploadService.setReturnTo({ url: ['/dev/bulk-upload/workplace-references'] });
-  }
-
-  public getFirstErrorMessage(item: string): string {
-    const errorType = Object.keys(this.form.get(item).errors)[0];
-    return this.errorSummaryService.getFormErrorMessage(item, errorType, this.formErrorsMap);
-  }
-
-  protected generateRequest(): LocalIdentifiersRequest {
-    return {
-      localIdentifiers: Object.keys(this.form.value).map((key) => ({
-        uid: key.replace('reference-', ''),
-        value: this.form.value[key],
-      })),
-    };
   }
 }
